backend/pyproject.toml
[project]
name = "nutrify-ai-backend"
version = "0.1.0"
description = "Nutrify-AI FastAPI Backend with LangChain Agents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Database
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "greenlet>=3.0.0",
    
    # Redis & Caching
    "redis>=5.0.1",
    "hiredis>=2.2.3",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    
    # LangChain & AI
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-community>=0.0.13",
    "langsmith>=0.0.69",
    "openai>=1.6.0",
    "tiktoken>=0.5.2",
    
    # Vector Store
    "faiss-cpu>=1.7.4",
    "pgvector>=0.2.4",
    
    # HTTP & External APIs
    "httpx>=0.25.2",
    "aiohttp>=3.9.1",
    
    # Background Tasks
    "celery>=5.3.4",
    "flower>=2.0.1",
    
    # Utilities
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    "email-validator>=2.1.0",
    
    # Monitoring & Logging
    "structlog>=24.1.0",
    "sentry-sdk[fastapi]>=1.39.0",
    
    # Payment Processing
    "stripe>=7.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.2",
    "faker>=21.0.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
